Rename byte to i8 test file
